@@ -1,92 +1,105 @@
 {
-  "type": "project",
-  "license": "proprietary",
-  "require": {
-    "php": "~7.4",
-    "ext-json": "*",
-    "ext-openssl": "*",
-    "ext-pgsql": "*",
-    "ext-sodium": "*",
-    "doctrine/common": "^2",
-    "doctrine/doctrine-bundle": "^2",
-    "doctrine/doctrine-migrations-bundle": "^2",
-    "doctrine/orm": "^2",
-    "jms/serializer-bundle": "^3.5",
-    "lexik/jwt-authentication-bundle": "^2.6",
-    "nelmio/api-doc-bundle": "^3.6",
-    "ramsey/uuid-doctrine": "^1.5",
-    "sensio/framework-extra-bundle": "^5.5",
-    "stof/doctrine-extensions-bundle": "^1.3",
-    "symfony/asset": "5.0.*",
-    "symfony/console": "5.0.*",
-    "symfony/dotenv": "5.0.*",
-    "symfony/flex": "^1.1",
-    "symfony/form": "5.0.*",
-    "symfony/framework-bundle": "5.0.*",
-    "symfony/security-bundle": "5.0.*",
-    "symfony/twig-bundle": "5.0.*",
-    "symfony/validator": "5.0.*",
-    "symfony/yaml": "5.0.*",
-    "twig/extra-bundle": "^2.12|^3.0",
-    "twig/twig": "^2.12|^3.0"
-  },
-  "config": {
-    "preferred-install": {
-      "*": "dist"
+    "type": "project",
+    "license": "proprietary",
+    "minimum-stability": "dev",
+    "prefer-stable": true,
+    "require": {
+        "php": ">=8.1",
+        "ext-ctype": "*",
+        "ext-iconv": "*",
+        "ext-openssl": "*",
+        "ext-pgsql": "*",
+        "doctrine/annotations": "^1.0",
+        "doctrine/doctrine-bundle": "^2.7",
+        "doctrine/doctrine-migrations-bundle": "^3.2",
+        "doctrine/orm": "^2.13",
+        "jms/serializer-bundle": "^3.5",
+        "lexik/jwt-authentication-bundle": "^2.6",
+        "nelmio/api-doc-bundle": "^4.0",
+        "phpdocumentor/reflection-docblock": "^5.3",
+        "phpstan/phpdoc-parser": "^1.15",
+        "ramsey/uuid-doctrine": "^1.5",
+        "stof/doctrine-extensions-bundle": "^1.3",
+        "symfony/asset": "6.2.*",
+        "symfony/console": "6.2.*",
+        "symfony/doctrine-messenger": "6.2.*",
+        "symfony/dotenv": "6.2.*",
+        "symfony/expression-language": "6.2.*",
+        "symfony/flex": "^2",
+        "symfony/form": "6.2.*",
+        "symfony/framework-bundle": "6.2.*",
+        "symfony/mailer": "6.2.*",
+        "symfony/mime": "6.2.*",
+        "symfony/monolog-bundle": "^3.0",
+        "symfony/notifier": "6.2.*",
+        "symfony/process": "6.2.*",
+        "symfony/property-access": "6.2.*",
+        "symfony/property-info": "6.2.*",
+        "symfony/runtime": "6.2.*",
+        "symfony/security-bundle": "6.2.*",
+        "symfony/translation": "6.2.*",
+        "symfony/twig-bundle": "6.2.*",
+        "symfony/validator": "6.2.*",
+        "symfony/yaml": "6.2.*",
+        "twig/extra-bundle": "^2.12|^3.0",
+        "twig/twig": "^2.12|^3.0"
     },
-    "sort-packages": true,
-    "allow-plugins": {
-      "symfony/flex": true
-    }
-  },
-  "autoload": {
-    "psr-4": {
-      "App\\": "src/"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "Tests\\": "tests/"
-    }
-  },
-  "replace": {
-    "paragonie/random_compat": "2.*",
-    "symfony/polyfill-ctype": "*",
-    "symfony/polyfill-iconv": "*",
-    "symfony/polyfill-php71": "*",
-    "symfony/polyfill-php70": "*",
-    "symfony/polyfill-php56": "*"
-  },
-  "scripts": {
-    "auto-scripts": {
-      "cache:clear": "symfony-cmd",
-      "assets:install %PUBLIC_DIR%": "symfony-cmd"
+    "config": {
+        "allow-plugins": {
+            "symfony/flex": true,
+            "symfony/runtime": true
+        },
+        "sort-packages": true
     },
-    "post-install-cmd": [
-      "@auto-scripts"
-    ],
-    "post-update-cmd": [
-      "@auto-scripts"
-    ]
-  },
-  "conflict": {
-    "symfony/symfony": "*"
-  },
-  "extra": {
-    "symfony": {
-      "allow-contrib": false,
-      "require": "5.0.*"
+    "autoload": {
+        "psr-4": {
+            "App\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "App\\Tests\\": "tests/"
+        }
+    },
+    "replace": {
+        "symfony/polyfill-ctype": "*",
+        "symfony/polyfill-iconv": "*",
+        "symfony/polyfill-php72": "*",
+        "symfony/polyfill-php73": "*",
+        "symfony/polyfill-php74": "*",
+        "symfony/polyfill-php80": "*",
+        "symfony/polyfill-php81": "*"
+    },
+    "scripts": {
+        "auto-scripts": {
+            "cache:clear": "symfony-cmd",
+            "assets:install %PUBLIC_DIR%": "symfony-cmd"
+        },
+        "post-install-cmd": [
+            "@auto-scripts"
+        ],
+        "post-update-cmd": [
+            "@auto-scripts"
+        ]
+    },
+    "conflict": {
+        "symfony/symfony": "*"
+    },
+    "extra": {
+        "symfony": {
+            "allow-contrib": false,
+            "require": "6.2.*"
+        }
+    },
+    "require-dev": {
+        "doctrine/doctrine-fixtures-bundle": "^3.4",
+        "phpunit/phpunit": "^9.5",
+        "symfony/browser-kit": "6.2.*",
+        "symfony/css-selector": "6.2.*",
+        "symfony/debug-bundle": "6.2.*",
+        "symfony/maker-bundle": "^1.0",
+        "symfony/phpunit-bridge": "^6.2",
+        "symfony/stopwatch": "6.2.*",
+        "symfony/web-profiler-bundle": "6.2.*"
     }
-  },
-  "require-dev": {
-    "dama/doctrine-test-bundle": "^6.3",
-    "doctrine/doctrine-fixtures-bundle": "^3.3",
-    "phpunit/phpunit": "^9.5",
-    "symfony/browser-kit": "5.0.*",
-    "symfony/css-selector": "5.0.*",
-    "symfony/maker-bundle": "^1.15",
-    "symfony/phpunit-bridge": "^5.0",
-    "symfony/stopwatch": "5.0.*",
-    "symfony/web-profiler-bundle": "5.0.*"
-  }
 }
