@@ -6,11 +6,10 @@ namespace App\Controller;
 
 use RuntimeException;
 use App\Model\LoginModel;
-use Swagger\Annotations as SWG;
+use OpenApi\Annotations as OA;
 use JMS\Serializer\SerializerInterface;
 use Nelmio\ApiDocBundle\Annotation\Model;
 use App\Controller\Traits\SerializerAware;
-use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
@@ -20,8 +19,6 @@ class LoginController extends AbstractController
 
     /**
      * ConnectionController constructor.
-     *
-     * @param SerializerInterface $serializer
      */
     public function __construct(SerializerInterface $serializer)
     {
@@ -30,24 +27,19 @@ class LoginController extends AbstractController
 
     /**
      * Initialise sessions with encryption API (Token valid for 30s)
-     * @Route(path="/login", methods={"POST"}, name="app_login")
-     * @SWG\Tag(name="Authentification")
-     * @SWG\Parameter(
-     *     name="body",
-     *     in="body",
-     *     type="json",
-     *     @SWG\Schema(@Model(type=LoginModel::class))
-     * )
-     * @SWG\Response(
+     * @OA\Tag(name="Authentification")
+     * @OA\RequestBody(@Model(type=LoginModel::class))
+     * @OA\Response(
      *     response=200,
      *     description="User connected",
-     *     @SWG\Schema(
+     *     @OA\Schema(
      *        type="object",
      *        example={"token": "gjc7834ace3-8525-4814-bf0f-b7146bc9e8ab"}
      *     )
      * )
      */
-    final public function login(): Response
+    #[Route(path: '/login', name: 'app_login', methods: ['post'])]
+    final public function login(): never
     {
         throw new RuntimeException(
             'You may have screwed the firewall configuration, this function should not have been called.'
