@@ -12,20 +12,17 @@ use Doctrine\Bundle\FixturesBundle\Fixture;
 
 class UserFixtures extends Fixture
 {
-    public const REFERENCE_ADMIN = 'user-admin';
-    public const REFERENCE_USER  = 'user-user';
+    final public const REFERENCE_ADMIN = 'user-admin';
+    final public const REFERENCE_USER  = 'user-user';
 
-    private UserService $userService;
-
-    public function __construct(UserService $userService)
+    public function __construct(private readonly UserService $userService)
     {
-        $this->userService = $userService;
     }
 
     public function load(ObjectManager $manager): void
     {
-        $admin = $this->userService->createUser('default', '766af668-0c19-4624-bcb4-bdb09ce4dada');
-        $user  = $this->userService->createUser('default', '0b819649-bef4-4fb9-a6b4-7b7b0b69961c');
+        $admin = $this->userService->createUser('admin-passwd', '766af668-0c19-4624-bcb4-bdb09ce4dada');
+        $user  = $this->userService->createUser('user-passwd', '0b819649-bef4-4fb9-a6b4-7b7b0b69961c');
 
         $this->setReference(self::REFERENCE_ADMIN, $admin);
         $this->setReference(self::REFERENCE_USER, $user);
