@@ -1,42 +1,31 @@
 <?php
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace App\Service;
 
+use App\Entity\Media;
 use Exception;
 use Generator;
-use App\Entity\Media;
-use RuntimeException;
 use Ramsey\Uuid\Uuid;
+use RuntimeException;
 use Symfony\Component\Filesystem\Filesystem;
-use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\HttpFoundation\File\Exception\FileNotFoundException;
+use Symfony\Component\HttpFoundation\File\File;
 
 class MediaService
 {
-    const SECRET_KEY_LENGHT = 256;
+    final const SECRET_KEY_LENGHT = 256;
 
-    private string                $mediaFolder;
-
-    private Filesystem            $filesystem;
-
-    private FileEncryptionService $fileEncryption;
-
-    /**
-     * @var EncryptionService
-     */
-    private EncryptionService $encryptionService;
+    private readonly Filesystem $filesystem;
 
     public function __construct(
-        EncryptionService $encryptionService,
-        FileEncryptionService $fileEncryption,
-        string $mediaFolder
-    ) {
-        $this->encryptionService = $encryptionService;
-        $this->mediaFolder       = $mediaFolder;
-        $this->fileEncryption    = $fileEncryption;
-        $this->filesystem        = new Filesystem();
+        private readonly EncryptionService     $encryptionService,
+        private readonly FileEncryptionService $fileEncryption,
+        private readonly string                $mediaFolder
+    )
+    {
+        $this->filesystem = new Filesystem();
     }
 
     /**
@@ -50,7 +39,7 @@ class MediaService
             return null;
         }
 
-        $password          = $this->generatePassword(self::SECRET_KEY_LENGHT);
+        $password = $this->generatePassword(self::SECRET_KEY_LENGHT);
         $encryptedPassword = $this->encryptionService->encryptData($password);
 
         $media
@@ -114,12 +103,12 @@ class MediaService
     private function generatePassword(int $length): string
     {
         $chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz' .
-                 '0123456789-=~!@#$%^&*()_+,.<>?;:[]{}';
+            '0123456789-=~!@#$%^&*()_+,.<>?;:[]{}';
 
         $password = '';
-        $max      = strlen($chars) - 1;
+        $max = strlen($chars) - 1;
 
-        for ($index = 0; $index < $length; $index ++) {
+        for ($index = 0; $index < $length; $index++) {
             $password .= $chars[random_int(0, $max)];
         }
 
