@@ -1,15 +1,13 @@
 <?php
 
-namespace Tests\Controller;
+namespace App\Tests\Controller;
 
-use Swift_Message;
-use Tests\ApiTestCase;
-use App\Model\RegisterModel;
+use App\Tests\ApiTestCase;
 use Symfony\Component\HttpFoundation\Response;
 
 class LoginControllerTest extends ApiTestCase
 {
-    public function testLoginBadCredencial(): void
+    public function testLoginBadUuid(): void
     {
         $client = static::createClient();
 
@@ -20,6 +18,26 @@ class LoginControllerTest extends ApiTestCase
             [],
             ['CONTENT_TYPE' => 'application/json'],
             '{"username":"baduser","password":"badpassword"}'
+        );
+
+        $this->assertEquals(
+            Response::HTTP_INTERNAL_SERVER_ERROR,
+            $client->getResponse()->getStatusCode(),
+            'can use a bad Uuid as username'
+        );
+    }
+
+    public function testLoginBadCredencial(): void
+    {
+        $client = static::createClient();
+
+        $client->request(
+            'POST',
+            '/login',
+            [],
+            [],
+            ['CONTENT_TYPE' => 'application/json'],
+            '{"username":"766af668-0c19-4624-bcb4-bdb09ce4dada","password":"badpassword"}'
         );
 
         $this->assertEquals(
@@ -39,7 +57,7 @@ class LoginControllerTest extends ApiTestCase
             [],
             [],
             ['CONTENT_TYPE' => 'application/json'],
-            '{"username":"766af668-0c19-4624-bcb4-bdb09ce4dada","password":"default"}'
+            '{"username":"766af668-0c19-4624-bcb4-bdb09ce4dada","password":"admin-passwd"}'
         );
 
         $this->assertEquals(
@@ -59,7 +77,7 @@ class LoginControllerTest extends ApiTestCase
             [],
             [],
             ['CONTENT_TYPE' => 'application/json'],
-            '{"username":"0b819649-bef4-4fb9-a6b4-7b7b0b69961c","password":"default"}'
+            '{"username":"0b819649-bef4-4fb9-a6b4-7b7b0b69961c","password":"user-passwd"}'
         );
 
         $this->assertEquals(
